requirements.txt
Flask==2.3.0
mysql-connector-python==8.0.33
scikit-learn==1.3.0
pandas==2.0.0
numpy==1.24.0
bcrypt==4.0.1
joblib==1.3.0